+++ Runtime/OculusLoader.cs
@@ -138,6 +138,7 @@ namespace Unity.XR.Oculus
                 userDefinedSettings.protectedContext = (ushort)(settings.ProtectedContext ? 1 : 0);
                 userDefinedSettings.focusAware = (ushort)(settings.FocusAware ? 1 : 0);
                 userDefinedSettings.optimizeBufferDiscards = (ushort)(settings.OptimizeBufferDiscards ? 1 : 0);
+                userDefinedSettings.phaseSync = (ushort)(settings.PhaseSync ? 1 : 0);
                 NativeMethods.SetUserDefinedSettings(userDefinedSettings);
             }